imageGenerationWithDiffusionModels

Documentation for imageGenerationWithDiffusionModels.

This Julia package implements a diffusion model to generate images of digits. It learns how noise alters the data and then predicts how to reverse this process to retrieve digits out of pure noise. This project can serve as a basis for more complex diffusion models.

imageGenerationWithDiffusionModels.Embeddings.LearnedTEmbedding

imageGenerationWithDiffusionModels.FeatureEncoderNetwork._add_unet_level

imageGenerationWithDiffusionModels.FeatureEncoderNetwork.make_down_path

imageGenerationWithDiffusionModels.Scheduler.cosine_beta_schedule

imageGenerationWithDiffusionModels.Training.train

imageGenerationWithDiffusionModels.UNet.make_unet

imageGenerationWithDiffusionModels.add_noise_to_image

imageGenerationWithDiffusionModels.add_noise_to_image_visualization

imageGenerationWithDiffusionModels.load_digits_data

imageGenerationWithDiffusionModels.visualize_noising_of_image

imageGenerationWithDiffusionModels.add_noise_to_image — Function

add_noise_to_image(img::Vector{Float64}, noise_step::Int64, alpha_bar::Vector{Float64}; rng = Random.GLOBAL_RNG)

Applies Gaussian noise to an image.

Arguments

img::Matrix{Float32} : Input image

noise_step::Int64 : A noising step

alpha_bar::Vector{Float64} : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng::Random.TaskLocalRNG: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

A noised version of image and the applied noise

imageGenerationWithDiffusionModels.add_noise_to_image_visualization — Function

add_noise_to_image_visualization(img::Vector{Float64}, noise_step::Int64, alpha_bar::Vector{Float64}; rng = Random.GLOBAL_RNG)

Applies Gaussian noise to an image.

Arguments

img::Matrix{Float32} : Input image

noise_step::Int64 : A noising step

alpha_bar::Vector{Float64} : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng::Random.TaskLocalRNG: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

A noised version of image

imageGenerationWithDiffusionModels.load_digits_data — Method

load_digits_data(filepath::String)

Loads digits data from a .mat filepath.

Arguments

filepath::String: A filepath to .mat digits data.

...

imageGenerationWithDiffusionModels.visualize_noising_of_image — Function

visualize_noising_of_image(img, noise_step, alpha_bar, rng = Random.GLOBAL_RNG)

Visualizes the Gaussian noising process of an image.

Arguments

img::Matrix{Float32} : Input image

noise_step::StepRange{Int64, Int64} : Number of noising steps (in each step, noise is applied to the output of the previous step)

alpha_bar::Vector{Float64} : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

An image visualizing the Gaussian noising process of an image horizontally.

imageGenerationWithDiffusionModels.FeatureEncoderNetwork._add_unet_level — Method

_add_unet_level(
    in_out::Vector{Tuple{Int,Int}}, emb_dim::Int, level::Int;
    block_layer, num_blocks_per_level::Int, block_groups::Int, num_attention_heads::Int
)

This function adds a level to the UNet through the use of the ConditionalSkipConnection. The skipped layers are defined recursively until the intended level-count of the UNet has been reached. When the break condition is reached, the middle of the Unet is created

imageGenerationWithDiffusionModels.FeatureEncoderNetwork.make_down_path — Method

makedownpath(; channels=(64,128,256), embdim=128, inch=1)

Returns a tuple (encode, out_channels).

encode(img, t_emb) : (latent, intermediates)

out_channels :final channel count (needed by bottleneck)

TODO: Visualize the intermediates of MNIST

imageGenerationWithDiffusionModels.Scheduler.cosine_beta_schedule — Function

cosine_beta_schedule(num_timesteps::Int, s=0.008)

Generates a noise schedule based on a cosine beta function.

Arguments

num_timesteps::Int : Number of timesteps.

s : A variable, which varies the shape of the schedule. Defaults to: 0.008.

Returns

A beta schedule.

imageGenerationWithDiffusionModels.Embeddings.LearnedTEmbedding — Method

LearnedTEmbedding(dim; inner = 4*dim)

Two-layer MLP that maps a scalar time-step to a length-dim embedding vector.

imageGenerationWithDiffusionModels.UNet.make_unet — Method

Dummy Unet model

Down-path : built by make_down_path Bottleneck : #TODO Up-path : #TODO

imageGenerationWithDiffusionModels.Training.train — Method

train(;FILE_PATH::String = "./example/SyntheticImages500.mat",
num_timesteps::Int = 100,
learning_rate::Real = 0.0001,
epochs::Int = 15,
batch_size::Int = 32,
shuffle::Bool = true,
model::unet = unet(
    1,
    5,
    16,
    LearnedTEmbedding(128),
    128;
    num_blocks_per_level=1
))

Trains a diffusion model on a digit dataset.

Arguments

FILE_PATH::String : A filepath to .mat digits data. Defaults to "./example/SyntheticImages500.mat".

num_timesteps::Int : Number of noising steps. Defaults to 100.

cosine::Bool : Cosine beta scheduling. Otherwise linear. Defaults to true.

learning_rate::Real : A training learning rate. Defaults to 0.0001.

epochs::Int : Number of training epochs. Defaults to 15.

batch_size::Int : Size of training batches. Defaults to 32.

shuffle::Bool : Shuffle training data. Defaults to true.

model::unet : Diffusion model. Defaults to unet(1, 5, 16, LearnedTEmbedding(128), 128; num_blocks_per_level=1).

Returns

A trained diffusion model. Saves the model as "model.bson".