imageGenerationWithDiffusionModels

Documentation for imageGenerationWithDiffusionModels.

This Julia package implements a diffusion model to generate images of digits. It learns how noise alters the data and then predicts how to reverse this process to retrieve digits out of pure noise. This project can serve as a basis for more complex diffusion models.

imageGenerationWithDiffusionModels.Blocks.ConditionalChain

imageGenerationWithDiffusionModels.Blocks.ConditionalChain

imageGenerationWithDiffusionModels.Blocks.ConditionalChain

imageGenerationWithDiffusionModels.Blocks.ConditionalChain

imageGenerationWithDiffusionModels.Blocks.ConditionalSkipConnection

imageGenerationWithDiffusionModels.Blocks.ConditionalSkipConnection

imageGenerationWithDiffusionModels.Blocks.TResBlock

imageGenerationWithDiffusionModels.Embeddings.LearnedTEmbedding

imageGenerationWithDiffusionModels.UNet.unet

imageGenerationWithDiffusionModels.UNet.unet

imageGenerationWithDiffusionModels.UNet.unet

imageGenerationWithDiffusionModels.Blocks.Downsample

imageGenerationWithDiffusionModels.Blocks.Upsampling

imageGenerationWithDiffusionModels.FeatureEncoderNetwork._add_unet_level

imageGenerationWithDiffusionModels.FeatureEncoderNetwork.make_down_path

imageGenerationWithDiffusionModels.Scheduler.cosine_beta_schedule

imageGenerationWithDiffusionModels.Training.train

imageGenerationWithDiffusionModels.add_noise_to_image

imageGenerationWithDiffusionModels.add_noise_to_image_visualization

imageGenerationWithDiffusionModels.load_digits_data

imageGenerationWithDiffusionModels.visualize_noising_of_image

imageGenerationWithDiffusionModels.add_noise_to_image — Function

add_noise_to_image(img::AbstractArray, noise_step::Int, alpha_bar::AbstractVector, rng = Random.GLOBAL_RNG)

Applies Gaussian noise to an image and returns it.

Arguments

img::AbstractArray : Input image

noise_step::Int : A noising step

alpha_bar::AbstractVector : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng::Random.TaskLocalRNG: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

A noised version of image in form of a matrix and the applied noise.

imageGenerationWithDiffusionModels.add_noise_to_image_visualization — Function

add_noise_to_image_visualization(img, noise_step::Int64, alpha_bar::Vector{Float64}; rng = Random.GLOBAL_RNG)

Applies Gaussian noise to an image. Utility function purely for visualization purposes.

Arguments

img::Matrix{Float32} : Input image

noise_step::Int64 : A noising step

alpha_bar::Vector{Float64} : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng::Random.TaskLocalRNG: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

A noised version of image in form of a matrix.

imageGenerationWithDiffusionModels.load_digits_data — Method

load_digits_data(filepath::String)

Loads digits data from a .mat filepath.

Arguments

filepath::String: A filepath to .mat digits data.

...

imageGenerationWithDiffusionModels.visualize_noising_of_image — Function

visualize_noising_of_image(img, noise_step, alpha_bar, rng = Random.GLOBAL_RNG)

Visualizes the Gaussian noising process of an image.

Arguments

img::Matrix{Float32} : Input image

noise_step::StepRange{Int64, Int64} : Number of noising steps (in each step, noise is applied to the output of the previous step)

alpha_bar::Vector{Float64} : Vector of noise parameters. Length must be at least "noise_step". Comupted by taking the Cumulative product of (1-"variance schedule")

rng: Random number generator. Defaults to: Random.GLOBAL_RNG.

Returns

An image visualizing the Gaussian noising process of an image horizontally.

imageGenerationWithDiffusionModels.FeatureEncoderNetwork._add_unet_level — Method

_add_unet_level(
    in_out::Vector{Tuple{Int,Int}}, emb_dim::Int, level::Int;
    block_layer, num_blocks_per_level::Int, block_groups::Int, num_attention_heads::Int
)

This function adds a level to the UNet through the use of the ConditionalSkipConnection. The skipped layers are defined recursively until the intended level-count of the UNet has been reached. When the break condition is reached, the middle of the Unet is created

imageGenerationWithDiffusionModels.FeatureEncoderNetwork.make_down_path — Method

make_down_path(; channels=(64,128,256), emb_dim=128, in_ch=1)

Returns a tuple (encode, out_channels).

encode(img, t_emb) : (latent, intermediates)

out_channels :final channel count (needed by bottleneck)

TODO: Visualize the intermediates of MNIST

imageGenerationWithDiffusionModels.Blocks.ConditionalChain — Type

ConditionalChain{T<:Union{Tuple,NamedTuple}} <: AbstractParallel

A Chain operator, that accepts multiple inputs (conditional inputs). See also Flux.Chain

imageGenerationWithDiffusionModels.Blocks.ConditionalChain — Method

ConditionalChain(xs...)

Constructs a ConditionalChain object from the layers xs...

imageGenerationWithDiffusionModels.Blocks.ConditionalChain — Method

(c::ConditionalChain)(x, ys...)

Forward operator for the ConditionalChain The Layers of the Chain are applied to either only x or both x and ys... depending on the number of inputs, that the respective layer accepts.

imageGenerationWithDiffusionModels.Blocks.ConditionalChain — Method

ConditionalChain(; kw...)

Keyword constructor for the ConditionalChain.

imageGenerationWithDiffusionModels.Blocks.ConditionalSkipConnection — Type

ConditionalSkipConnection{T,F} <: AbstractParallel

A skip connection, that accepts multiple inputs (conditional inputs).

The struct holds layers that are meant to be skipped and connection which rejoins the output of the skipped layers with the input. See also Flux.SkipConnection

imageGenerationWithDiffusionModels.Blocks.ConditionalSkipConnection — Method

(skip::ConditionalSkipConnection)(x, ys...)

Forward pass for the ConditionalSkipConnection.

layers is applied to xs and ys. The output is then fused with x by applying connection to them.

imageGenerationWithDiffusionModels.Blocks.TResBlock — Type

struct TResBlock<: AbstractParallel

A Residual block with multiple inputs (conditional inputs)

imageGenerationWithDiffusionModels.Blocks.Downsample — Method

Downsample()

Applies a 2x2 MaxPooling to the Input.

imageGenerationWithDiffusionModels.Blocks.Upsampling — Method

Upsampling(channels::Pair{<:Integer, <:Integer})

Applies a 2x2 Upsampling and a 3x3 padded Convolution parametrized by channels

imageGenerationWithDiffusionModels.Scheduler.cosine_beta_schedule — Function

cosine_beta_schedule(num_timesteps::Int, s=0.008)

Generates a noise schedule based on a cosine beta function.

Arguments

num_timesteps::Int : Number of timesteps.

s : A variable, which varies the shape of the schedule. Defaults to: 0.008.

Returns

A beta schedule.

imageGenerationWithDiffusionModels.Embeddings.LearnedTEmbedding — Method

LearnedTEmbedding(dim; inner = 4*dim)

Two-layer MLP that maps a scalar time-step to a length-dim embedding vector.

imageGenerationWithDiffusionModels.UNet.unet — Type

unet{E,C<:ConditionalChain}

Struct of the UNet Model.

imageGenerationWithDiffusionModels.UNet.unet — Method

(u::unet)(x::AbstractArray, timesteps::AbstractVector)

Forward pass for the UNet model.

Usually takes an Image in form of a Vector and a timestep as inputs and returns an Image in form of a vector.

imageGenerationWithDiffusionModels.UNet.unet — Method

unet(
    in_channels::Int,
    num_levels::Int,
    model_dim::Int,
    time_embed,
    emb_dim::Int;
    block_layer=TResBlock,
    num_blocks_per_level::Int=1,
)

Constructs a UNet Model.

The depth of the Model is parametrized by num_levels. time_embed is a time embedding block with an output length emb_dim.

imageGenerationWithDiffusionModels.Training.train — Method

train(;FILE_PATH::String = "./example/SyntheticImages500.mat",
num_timesteps::Int = 100,
learning_rate::Real = 0.0001,
epochs::Int = 15,
batch_size::Int = 32,
shuffle::Bool = true,
model::unet = unet(
    1,
    5,
    16,
    LearnedTEmbedding(128),
    128;
    num_blocks_per_level=1
))

Trains a diffusion model on a digit dataset.

Arguments

FILE_PATH::String : A filepath to .mat digits data. Defaults to "./example/SyntheticImages500.mat".

num_timesteps::Int : Number of noising steps. Defaults to 100.

cosine::Bool : Cosine beta scheduling. Otherwise linear. Defaults to true.

learning_rate::Real : A training learning rate. Defaults to 0.0001.

epochs::Int : Number of training epochs. Defaults to 15.

batch_size::Int : Size of training batches. Defaults to 32.

shuffle::Bool : Shuffle training data. Defaults to true.

model::unet : Diffusion model. Defaults to unet(1, 5, 16, LearnedTEmbedding(128), 128; num_blocks_per_level=1).

Returns

A trained diffusion model. Saves the model as "model.bson".